refactor rle module
