+++ encodings/fastlanes/src/rle/vtable/encode.rs
@@ -1,0 +1,20 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use vortex_array::Canonical;
+use vortex_array::vtable::EncodeVTable;
+use vortex_error::VortexResult;
+
+use super::{RLEEncoding, RLEVTable};
+use crate::RLEArray;
+
+impl EncodeVTable<RLEVTable> for RLEVTable {
+    fn encode(
+        _encoding: &RLEEncoding,
+        canonical: &Canonical,
+        _like: Option<&RLEArray>,
+    ) -> VortexResult<Option<RLEArray>> {
+        let array = canonical.clone().into_primitive();
+        Ok(Some(RLEArray::encode(&array)?))
+    }
+}
